add run/debug file command
