+++ packages/salesforcedx-vscode-lwc/package.json
@@ -124,6 +124,14 @@
     },
     "commands": [
       {
+        "command": "sfdx.force.lightning.lwc.test.file.run",
+        "title": "%force_lightning_lwc_test_file_run_text%"
+      },
+      {
+        "command": "sfdx.force.lightning.lwc.test.file.debug",
+        "title": "%force_lightning_lwc_test_file_debug_text%"
+      },
+      {
         "command": "sfdx.force.lightning.lwc.test.case.run",
         "title": "%force_lightning_lwc_test_case_run_text%",
         "icon": {
