+++ packages/salesforcedx-vscode-lwc/package.nls.json
@@ -1,5 +1,7 @@
 {
   "force_lightning_lwc_test_view_name": "LWC Tests",
-  "force_lightning_lwc_test_case_run_text": "SFDX: Invoke Lightning Web Component Test Case",
+  "force_lightning_lwc_test_file_run_text": "SFDX: Run Lightning Web Component Test File",
+  "force_lightning_lwc_test_file_debug_text": "SFDX: Debug Lightning Web Component Test File",
+  "force_lightning_lwc_test_case_run_text": "SFDX: Run Lightning Web Component Test Case",
   "force_lightning_lwc_test_case_debug_text": "SFDX: Debug Lightning Web Component Test Case"
 }
