+++ packages/salesforcedx-vscode-lwc/src/testSupport/testIndexer/index.ts
@@ -8,7 +8,12 @@ import { JestTotalResults, parse } from 'jest-editor-support';
 import { Indexer } from 'lightning-lsp-common';
 import * as vscode from 'vscode';
 
-import { TestExecutionInfo, TestResultStatus, TestType } from '../types';
+import {
+  LwcJestTestResults,
+  TestExecutionInfo,
+  TestResultStatus,
+  TestType
+} from '../types';
 import { LWC_TEST_GLOB_PATTERN } from '../types/constants';
 
 export class LWCTestIndexer implements Indexer {
@@ -66,48 +71,6 @@ export async function findTestInfoFromLwcJestTestFile(
     console.error(error);
     return [];
   }
-}
-
-export interface LwcJestTestResults {
-  numFailedTestSuites: number;
-  numFailedTests: number;
-  numPassedTestSuites: number;
-  numPassedTests: number;
-  numPendingTestSuites: number;
-  numPendingTests: number;
-  numRuntimeErrorTestSuites: number;
-  numTotalTestSuites: number;
-  numTotalTests: number;
-  testResults: LwcJestTestFileResult[];
-}
-
-type LwcJestTestResultStatus =
-  | 'passed'
-  | 'failed'
-  | 'pending'
-  | 'skipped'
-  | 'pending'
-  | 'todo'
-  | 'disabled';
-
-export interface LwcJestTestFileResult {
-  status: LwcJestTestResultStatus;
-  startTime: number;
-  endTime: number;
-  name: string;
-  assertionResults: LwcJestTestAssertionResult[];
-}
-
-export interface LwcJestTestAssertionResult {
-  status: LwcJestTestResultStatus;
-  title: string;
-  ancestorTitles: string[];
-  failureMessages: string[];
-  fullName: string;
-  location: {
-    column: number;
-    line: number;
-  };
 }
 
 export const onDidUpdateTestResultsIndex = new vscode.EventEmitter<undefined>();
